Move browser evaluation helper to core
